% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2.R
\name{ahern}
\alias{ahern}
\title{Function to implement exact single-stage phase II design a-la AHern}
\usage{
ahern(p0, p1, alpha = 0.05, beta = 0.2, sample_sizes = 1:2000)
}
\arguments{
\item{p0}{response probability used in the definition of the null hypothesis}

\item{p1}{desirable response probability at which the trial is powered}

\item{alpha}{type I error (default 0.05)}

\item{beta}{type II error  (default 0.2)}

\item{sample_sizes}{checked sample sizes, should be enough for most
situations}
}
\description{
Function to implement exact single-stage phase II design a-la AHern
}
\examples{
# simple example
ahern(p0 = 0.05, p1 = 0.1)

# A'Hern's number validation
probs <- seq(0.05, 0.95, by = 0.05)
analyses <- expand.grid(p0 = probs, p1 = probs,
                        alpha = c(0.05, 0.01), beta = c(0.2, 0.1))
select <- with(analyses, p1 > p0)
analyses <- analyses[select, ]
ord <- with(analyses, order(p0, p1, -alpha, -beta))
analyses <- analyses[ord, ]
rownames(analyses) <- NULL
res <- apply(analyses, 1, function(x)
    ahern(p0 = x[1], p1 = x[2], alpha = x[3], beta = x[4]))
res <- do.call(rbind, res)
rownames(res) <- NULL
res$pp <- with(res, paste(cutoffs, sample_size, sep = '/'))
res$sample_size <- NULL
res$cutoffs <- NULL
res_spl <- split(res, res$p0)
## sink('ahern_validation.txt')
## res_spl
## sink()

}
\references{
A'Hern, Sample size tables for exact single-stage phase
II designs, Statist. Med. 2001; 20:859–866
}
