% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_percentage.R
\name{cp_n}
\alias{cp_n}
\title{Clopper-Pearson confidence interval width}
\usage{
cp_n(width, alpha = 0.05, p = 0.5, sample_size = 1:2000)
}
\arguments{
\item{width}{confidence interval width}

\item{alpha}{alpha}

\item{p}{expected proportion}

\item{sample_size}{investigated sample size}
}
\description{
Clopper-Pearson confidence interval width for a given sample size
and, alpha and p
}
