% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_percentage.R
\name{cp_width}
\alias{cp_width}
\title{Clopper-Pearson confidence interval width}
\usage{
cp_width(n, alpha = 0.05, p = 0.5)
}
\arguments{
\item{n}{sample size}

\item{alpha}{alpha}

\item{p}{expected proportion}
}
\description{
Clopper-Pearson confidence interval width for a given sample size
and, alpha and p
}
