% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_percentage.R
\name{wilson_width}
\alias{wilson_width}
\title{Wilson confidence interval width}
\usage{
wilson_width(n, alpha = 0.05, p = 0.5)
}
\arguments{
\item{n}{sample size}

\item{alpha}{alpha}

\item{p}{expected proportion}
}
\description{
Wilson confidence interval width for a given sample size and, alpha and p
}
